refactor ANSII colors into it's own component
